Fixed test K of the Unofficial Instruction test
I made changes to the stack pointer in a previous commit as a means to prevent a stack pointer value of 00 in a test from overwriting the bottom of the stack. That modified the results of test K, as it expected the stack pointer to have the value F6. My solution was to force the stack pointer to a known value for the test.